refactor: 리뷰 페이지 코드 정리 및 컴포넌트 분리 작업 중
