+++ constants/stats.ts
@@ -1,46 +1,90 @@
 export const seasonMapping: { [key: string]: string } = {
-  allSeasons: '사계절',
-  spring: '봄',
-  summer: '여름',
-  autumn: '가을',
-  winter: '겨울',
+  ALLSEASONS: '사계절',
+  SPRING: '봄',
+  SUMMER: '여름',
+  AUTUMN: '가을',
+  WINTER: '겨울',
 };
 
 export const longevityMapping: { [key: string]: string } = {
-  fullday: '하루종일',
-  halfday: '반나절',
-  fourhours: '3-4시간',
-  onehour: '1시간',
+  FULLDAY: '하루종일',
+  HALFDAY: '반나절',
+  FOURHOURS: '3-4시간',
+  ONEHOUR: '1시간',
 };
 
 export const intensityMapping: { [key: string]: string } = {
-  intense: '아주진함',
-  strong: '진함',
-  medium: '보통',
-  weak: '약함',
+  INTENSE: '아주진함',
+  STRONG: '진함',
+  MEDIUM: '보통',
+  WEAK: '약함',
 };
 
 export const styleMapping: { [key: string]: string } = {
-  chic: '시크한',
-  mature: '성숙한',
-  luxurious: '고급스러운',
-  elegant: '우아한',
-  masculine: '남성적인',
-  comfortable: '편안한',
-  serene: '차분한',
-  light: '가벼운',
-  neutural: '중성적인',
-  friendly: '친근한',
-  clean: '깨끗한',
-  sensual: '관능적인',
-  delicate: '은은한',
-  lively: '활기찬',
-  lovely: '사랑스러운',
-  bright: '밝은',
-  radiant: '화사한',
-  feminine: '여성스러운',
-  innocent: '청순한',
-  weighty: '무게감 있는',
-  soft: '부드러운',
-  cozy: '포근한',
+  CHIC: '시크한',
+  MATURE: '성숙한',
+  LUXURIUS: '고급스러운',
+  ELEGANT: '우아한',
+  MASCULINE: '남성적인',
+  COMFORTABLE: '편안한',
+  SERENE: '차분한',
+  LIGHT: '가벼운',
+  NEUTRAL: '중성적인',
+  FRIENDLY: '친근한',
+  CLEAN: '깨끗한',
+  SENSUAL: '관능적인',
+  DELICATE: '은은한',
+  LIVELY: '활기찬',
+  LOVELY: '사랑스러운',
+  BRIGHT: '밝은',
+  RADIANT: '화사한',
+  FEMININE: '여성스러운',
+  INNOCENT: '청순한',
+  WEIGHTY: '무게감 있는',
+  SOFT: '부드러운',
+  COZY: '포근한',
+};
+
+export const seasonMappingR: { [key: string]: string } = {
+  봄: 'SPRING',
+  여름: 'SUMMER',
+  가을: 'AUTUMN',
+  겨울: 'WINTER',
+  사계절: 'ALL',
+};
+export const longevityMappingR: { [key: string]: string } = {
+  '1시간': 'ONEHOUR',
+  '3-4시간': 'FOURHOURS',
+  반나절: 'HALFDAY',
+  하루종일: 'FULLDAY',
+};
+export const intensityMappingR: { [key: string]: string } = {
+  약함: 'WEAK',
+  보통: 'MEDIUM',
+  진함: 'STRONG',
+  아주진함: 'INTENSE',
+};
+export const styleMappingR: { [key: string]: string } = {
+  시크한: 'CHIC',
+  성숙한: 'MATURE',
+  고급스러운: 'LUXURIOUS',
+  우아한: 'ELEGANT',
+  남성적인: 'MASCULINE',
+  편안한: 'COMFORTABLE',
+  차분한: 'SERENE',
+  가벼운: 'LIGHT',
+  중성적인: 'NEUTRAL',
+  친근한: 'FRIENDLY',
+  깨끗한: 'CLEAN',
+  관능적인: 'SENSUAL',
+  은은한: 'DELICATE',
+  활기찬: 'LIVELY',
+  사랑스러운: 'LOVELY',
+  밝은: 'BRIGHT',
+  화사한: 'RADIANT',
+  여성스러운: 'FEMININE',
+  청순한: 'INNOCENT',
+  무게감있는: 'WEIGHTY',
+  부드러운: 'SOFT',
+  포근한: 'COZY',
 };
